Refactor code and added tanker
