+++ godot/scenes/characters/enemies/tanker.tscn
@@ -1,0 +1,23 @@
+[gd_scene load_steps=5 format=3 uid="uid://nbb3v15k8tu3"]
+
+[ext_resource type="PackedScene" uid="uid://dgwblviegvs7j" path="res://scenes/characters/enemy.tscn" id="1_o34qx"]
+[ext_resource type="Script" path="res://scripts/Characters/Enemy/Tanker.cs" id="2_d470p"]
+[ext_resource type="SpriteFrames" uid="uid://cvykc2va34ku3" path="res://assets/sprite_frames/grunt.tres" id="3_ri0aq"]
+[ext_resource type="AnimationLibrary" uid="uid://bvr7oyj3gbxlm" path="res://assets/sprite_frames/animation_library/Enemy_AnimationLibrary_new.res" id="4_0smyj"]
+
+[node name="Tanker" instance=ExtResource("1_o34qx")]
+script = ExtResource("2_d470p")
+
+[node name="Sprite2D (Preview)" parent="." index="0"]
+visible = false
+
+[node name="AnimatedSprite2D" parent="." index="1"]
+position = Vector2(5, -20)
+scale = Vector2(3, 3)
+sprite_frames = ExtResource("3_ri0aq")
+animation = &"attack"
+
+[node name="AnimationPlayer" parent="." index="10"]
+libraries = {
+"Enemy_AnimationLibrary": ExtResource("4_0smyj")
+}
