+++ godot/scripts/Components/MobSpawnerComponent.cs
@@ -26,11 +26,16 @@ public partial class MobSpawnerComponent : Node2D
         for (int i = 0; i < 10; i++)
         {
             Vector2 randomSpawnPosition = mainNode.GetRandomOutOfViewportPosition();
-            LoggingUtils.Info($"{randomSpawnPosition}");
             var enemy = Utils.CreateDummy(randomSpawnPosition, Scenes.Grunt);
             enemy.OnCharacterDeadEvent += enemy.DestroyCharacter;
-
-            enemy.CharacterStatComponent.AddStat("Health", 100.0f);
         }
+
+        for (int i = 0; i < 2; i++)
+        {
+            Vector2 randomSpawnPosition = mainNode.GetRandomOutOfViewportPosition();
+            var enemy = Utils.CreateDummy(randomSpawnPosition, Scenes.Tanker);
+            enemy.OnCharacterDeadEvent += enemy.DestroyCharacter;
+        }
+
     }
 }
